REFACTORE: (MOISAM-214) 이벤트 리스너의 폴더 구조 변경 및 필요없는 의존성 제거
