+++ src/main/java/com/meetup/server/event/domain/listener/StartPointChangedEvent.java
@@ -1,4 +1,4 @@
-package com.meetup.server.event.domain.value;
+package com.meetup.server.event.domain.listener;
 
 import java.util.UUID;
