+++ src/main/java/com/meetup/server/global/support/PerformanceAspect.java
@@ -1,17 +1,15 @@
 package com.meetup.server.global.support;
 
+import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 @Aspect
 @Component
+@Slf4j
 public class PerformanceAspect {
-
-    private static final Logger log = LoggerFactory.getLogger(PerformanceAspect.class);
 
     @Around("@annotation(com.meetup.server.global.support.Performance)")
     public Object measureExecutionTime(ProceedingJoinPoint joinPoint) throws Throwable {
